Refacto workflow builder
