+++ packages/twenty-server/src/engine/core-modules/workflow/resolvers/workflow-version-step.resolver.ts
@@ -1,22 +1,18 @@
-import { UseFilters, UseGuards } from '@nestjs/common';
+import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
 
-import { CreateDraftFromWorkflowVersionInput } from 'src/engine/core-modules/workflow/dtos/create-draft-from-workflow-version-input';
 import { CreateWorkflowVersionStepInput } from 'src/engine/core-modules/workflow/dtos/create-workflow-version-step-input.dto';
 import { DeleteWorkflowVersionStepInput } from 'src/engine/core-modules/workflow/dtos/delete-workflow-version-step-input.dto';
 import { UpdateWorkflowVersionStepInput } from 'src/engine/core-modules/workflow/dtos/update-workflow-version-step-input.dto';
 import { WorkflowActionDTO } from 'src/engine/core-modules/workflow/dtos/workflow-step.dto';
-import { WorkflowTriggerGraphqlApiExceptionFilter } from 'src/engine/core-modules/workflow/filters/workflow-trigger-graphql-api-exception.filter';
 import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
 import { AuthWorkspace } from 'src/engine/decorators/auth/auth-workspace.decorator';
 import { UserAuthGuard } from 'src/engine/guards/user-auth.guard';
 import { WorkspaceAuthGuard } from 'src/engine/guards/workspace-auth.guard';
-import { WorkflowVersionStepWorkspaceService } from 'src/modules/workflow/common/workspace-services/workflow-version-step.workspace-service';
-import { WorkflowVersionDTO } from 'src/engine/core-modules/workflow/dtos/workflow-version.dto';
+import { WorkflowVersionStepWorkspaceService } from 'src/modules/workflow/workflow-builder/workflow-version/workflow-step/workflow-version-step.workspace-service';
 
 @Resolver()
 @UseGuards(WorkspaceAuthGuard, UserAuthGuard)
-@UseFilters(WorkflowTriggerGraphqlApiExceptionFilter)
 export class WorkflowVersionStepResolver {
   constructor(
     private readonly workflowVersionStepWorkspaceService: WorkflowVersionStepWorkspaceService,
@@ -59,25 +55,5 @@ export class WorkflowVersionStepResolver {
       workflowVersionId,
       stepId,
     });
-  }
-
-  @Mutation(() => WorkflowVersionDTO)
-  async createDraftFromWorkflowVersion(
-    @AuthWorkspace() { id: workspaceId }: Workspace,
-    @Args('input')
-    {
-      workflowId,
-      workflowVersionIdToCopy,
-    }: CreateDraftFromWorkflowVersionInput,
-  ): Promise<WorkflowVersionDTO> {
-    return {
-      id: await this.workflowVersionStepWorkspaceService.createDraftFromWorkflowVersion(
-        {
-          workspaceId,
-          workflowId,
-          workflowVersionIdToCopy,
-        },
-      ),
-    };
   }
 }
