+++ packages/twenty-server/src/engine/core-modules/workflow/workflow-api.module.ts
@@ -1,18 +1,26 @@
 import { Module } from '@nestjs/common';
 
-import { WorkflowTriggerResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-trigger.resolver';
 import { WorkflowBuilderResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-builder.resolver';
-import { WorkflowTriggerModule } from 'src/modules/workflow/workflow-trigger/workflow-trigger.module';
+import { WorkflowTriggerResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-trigger.resolver';
 import { WorkflowVersionStepResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-version-step.resolver';
-import { WorkflowBuilderModule } from 'src/modules/workflow/workflow-builder/workflow-builder.module';
+import { WorkflowVersionResolver } from 'src/engine/core-modules/workflow/resolvers/workflow-version.resolver';
 import { WorkflowCommonModule } from 'src/modules/workflow/common/workflow-common.module';
+import { WorkflowBuilderModule } from 'src/modules/workflow/workflow-builder/workflow-builder.module';
+import { WorkflowVersionModule } from 'src/modules/workflow/workflow-builder/workflow-version/workflow-version.module';
+import { WorkflowTriggerModule } from 'src/modules/workflow/workflow-trigger/workflow-trigger.module';
 
 @Module({
-  imports: [WorkflowTriggerModule, WorkflowBuilderModule, WorkflowCommonModule],
+  imports: [
+    WorkflowTriggerModule,
+    WorkflowBuilderModule,
+    WorkflowCommonModule,
+    WorkflowVersionModule,
+  ],
   providers: [
     WorkflowTriggerResolver,
     WorkflowBuilderResolver,
     WorkflowVersionStepResolver,
+    WorkflowVersionResolver,
   ],
 })
 export class WorkflowApiModule {}
