+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-schema/types/input-schema.type.ts
