+++ packages/twenty-server/src/modules/workflow/workflow-builder/workflow-version/workflow-step/workflow-version-step.module.ts
@@ -3,21 +3,17 @@ import { Module } from '@nestjs/common';
 import { NestjsQueryTypeOrmModule } from '@ptc-org/nestjs-query-typeorm';
 
 import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
+import { ServerlessFunctionModule } from 'src/engine/metadata-modules/serverless-function/serverless-function.module';
 import { WorkflowSchemaModule } from 'src/modules/workflow/workflow-builder/workflow-schema/workflow-schema.module';
-import { WorkflowVersionStepModule } from 'src/modules/workflow/workflow-builder/workflow-version/workflow-step/workflow-version-step.module';
-import { WorkflowVersionModule } from 'src/modules/workflow/workflow-builder/workflow-version/workflow-version.module';
+import { WorkflowVersionStepWorkspaceService } from 'src/modules/workflow/workflow-builder/workflow-version/workflow-step/workflow-version-step.workspace-service';
 
 @Module({
   imports: [
     WorkflowSchemaModule,
-    WorkflowVersionModule,
-    WorkflowVersionStepModule,
+    ServerlessFunctionModule,
     NestjsQueryTypeOrmModule.forFeature([ObjectMetadataEntity], 'metadata'),
   ],
-  exports: [
-    WorkflowSchemaModule,
-    WorkflowVersionModule,
-    WorkflowVersionStepModule,
-  ],
+  providers: [WorkflowVersionStepWorkspaceService],
+  exports: [WorkflowVersionStepWorkspaceService],
 })
-export class WorkflowBuilderModule {}
+export class WorkflowVersionStepModule {}
